add check for wget

+++ packages.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
+if [ -e /usr/bin/wget ]; then 
+  exit 0; 
+fi
+
 yum -y install wget
 
 wget https://raw.githubusercontent.com/thomaskise/hello-nti-300/master/packages.txt
